requirements.txt
pip==20.1.1
Babel==2.8.0
chardet==3.0.4
DateTimeRange==1.0.0
Django==2.2.12
django-generate-secret-key==1.0.2
django-address==0.2.2
django-crispy-forms==1.9.1
django-phonenumber-field==4.0.0
importlib-metadata==1.6.0
mbstrdecoder==1.0.0
phonenumbers==8.12.3
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
typepy==1.1.1
zipp==1.2.0